requirements.txt
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
curl_cffi==0.13.0
Flask==3.1.2
frozendict==2.4.6
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
MarkupSafe==3.0.3
multitasking==0.0.12
numpy==2.3.4
packaging==25.0
pandas==2.3.3
peewee==3.18.2
platformdirs==4.5.0
protobuf==6.33.0
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.66
